freshfresh/requirements.txt
python-telegram-bot==20.7
web3==6.11.4
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
ccxt==4.1.77
pandas==2.1.4
numpy==1.24.3
cryptography==41.0.8
eth-account==0.9.0
